% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NO.NA.R
\name{NO.NA}
\alias{NO.NA}
\title{Getting rid of NA values}
\usage{
NO.NA(data)
}
\arguments{
\item{data}{is a data frame}
}
\value{
product is the data frame without NA's
}
\description{
Getting rid of NA values
}
