% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointPlot.R
\name{Pointplot}
\alias{Pointplot}
\title{Point plot using data with out NA values}
\usage{
Pointplot(data)
}
\arguments{
\item{data}{is the data frame without NA's}
}
\value{
product is a point plot with x as year and y as weight
}
\description{
Point plot using data with out NA values
}
