% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistical_test.R
\name{ST}
\alias{ST}
\title{Statistical test using data with out NA values}
\usage{
ST(data, weight, year)
}
\arguments{
\item{data}{is the data frame without NA's}

\item{weight}{is a column in the data frame}

\item{year}{is a column in the data frame}
}
\value{
product is a summary of model_fit
}
\description{
Statistical test using data with out NA values
}
